third_party/blink/web_tests/platform/generic/fast/forms/ValidityState-tooLong-input-expected.txt

Tests for tooLong flag with <input> elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

No maxlength and no value
PASS input.validity.tooLong is false


Non-dirty value
PASS input.value.length is 5
PASS input.validity.tooLong is false
PASS input.value.length is 6
PASS input.validity.tooLong is false


Dirty value and longer than maxLength
PASS input.value.length is 4
PASS input.validity.tooLong is true
PASS input.validity.tooLong is false


Sets a value via DOM property
PASS input.validity.tooLong is false


Disabling makes the control valid
PASS input.validity.tooLong is true
PASS input.disabled = true; input.validity.tooLong is true
PASS input.disabled = false; input.validity.tooLong is true


Grapheme length is not greater than maxLength though character length is greater
PASS input.validity.tooLong is false
PASS input.value.length is 3
PASS input.validity.tooLong is true


Change the type with a too long value
PASS input.validity.tooLong is true
PASS input.type = "number"; input.validity.tooLong is false


minlength and maxlength together
PASS input.validity.tooLong is true
PASS input.validity.tooShort is false
PASS input.validity.tooLong is false
PASS input.validity.tooShort is false
PASS input.validity.tooLong is false
PASS input.validity.tooShort is true


minlength and maxlength clashing
PASS input.validity.tooLong is true
PASS input.validity.tooShort is false
PASS input.validity.tooLong is true
PASS input.validity.tooShort is true
PASS input.validity.tooLong is false
PASS input.validity.tooShort is true
PASS input.validity.tooLong is false
PASS input.validity.tooShort is false
PASS successfullyParsed is true

TEST COMPLETE
